Move polygon validation methods to own class
